+++ lib/ui.dart
@@ -29,6 +29,7 @@ part 'src/ui/QuestionBody/RPUIDateTimeQuestionBody.dart';
 part 'src/ui/QuestionBody/RPUIImageChoiceQuestionBody.dart';
 part 'src/ui/QuestionBody/RPUIIntegerQuestionBody.dart';
 part 'src/ui/QuestionBody/RPUISliderQuestionBody.dart';
+part 'src/ui/QuestionBody/RPUITextInputQuestionBody.dart';
 part 'src/ui/RPStyles.dart';
 part 'src/ui/RPUICompletionStep.dart';
 part 'src/ui/RPUIConsentReviewStep.dart';